Task1_Create_a_Database_in_SQL_and_Mongo/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23 